+++ ndbc-core/src/main/java/io/trane/ndbc/datasource/Connection.java
@@ -15,11 +15,11 @@ public interface Connection {
 
   Future<List<Row>> query(String query);
 
-  Future<Integer> execute(String query);
+  Future<Long> execute(String query);
 
   Future<List<Row>> query(PreparedStatement query);
 
-  Future<Integer> execute(PreparedStatement query);
+  Future<Long> execute(PreparedStatement query);
 
   <R> Future<R> withTransaction(final Supplier<Future<R>> sup);
 }
