+++ ndbc-postgres/src/main/java/io/trane/ndbc/postgres/proto/SimpleExecuteExchange.java
@@ -9,15 +9,15 @@ import io.trane.ndbc.proto.Exchange;
 import io.trane.ndbc.proto.ServerMessage;
 import io.trane.ndbc.util.PartialFunction;
 
-public final class SimpleExecuteExchange implements Function<String, Exchange<Integer>> {
+public final class SimpleExecuteExchange implements Function<String, Exchange<Long>> {
 
   @Override
-  public final Exchange<Integer> apply(final String command) {
+  public final Exchange<Long> apply(final String command) {
     return Exchange.send(new Query(command))
         .thenReceive(commandComplete)
         .thenReceive(ReadyForQuery.class);
   }
 
-  private final PartialFunction<ServerMessage, Exchange<Integer>> commandComplete = PartialFunction
+  private final PartialFunction<ServerMessage, Exchange<Long>> commandComplete = PartialFunction
       .when(CommandComplete.class, msg -> Exchange.value(msg.rows));
 }
